Adapted LiteralTypesHighlightingTest to new version selection test
 #SCL-15634
